evc/demo_apps/chatbot/requirements.txt
gradio_tools
langchain_community
langchain
openai
gpt4all